+++ examples/peterdejong.cpp
@@ -96,7 +96,7 @@ int main(void) {
     theRamCanvas.writeRAWfile("peterdejong_" + std::to_string(j) + ".mrw");
 
     // Root image transform
-    theRamCanvas.applyHomoPixTfrm(&mjr::color1c16b::tfrmStdPow, static_cast<float>(1.0/p));
+    theRamCanvas.applyHomoPixTfrm(&mjr::color1c16b::tfrmStdPow, 1.0/p);
     maxII = static_cast<uint64_t>(65535.0 * pow(static_cast<double>(maxII)/65535.0, 1/p));
 
     // Log image transform
